Add files via upload
This upload is the second part of the SDPA_project. Since we did not pay attention to the need to upload GitHub before, this upload has completed most of the part 2 parts, including data preparation and cleaning, data analysis, and the still unfinished parts include two additional questions and summaries.